# docs/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Patchman-engine API",
        "description": "API of the Patch application on [console.redhat.com](https://console.redhat.com)\n\nSyntax of the `filter[name]` query parameters is described in  [Filters documentation](https://github.com/RedHatInsights/patchman-engine/wiki/API-custom-filters)",
        "contact": {},
        "license": {
            "name": "GPLv3",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "v1.18.52"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/api/patch/v1/export/systems": {
            "get": {
                "summary": "Export systems for my account",
                "description": "Export systems for my account",
                "operationId": "exportSystems",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Find matching text",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[id]",
                        "in": "query",
                        "description": "Filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[display_name]",
                        "in": "query",
                        "description": "Filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[last_evaluation]",
                        "in": "query",
                        "description": "Filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[last_upload]",
                        "in": "query",
                        "description": "Filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[rhsa_count]",
                        "in": "query",
                        "description": "Filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[rhba_count]",
                        "in": "query",
                        "description": "Filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[rhea_count]",
                        "in": "query",
                        "description": "Filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[other_count]",
                        "in": "query",
                        "description": "Filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[stale]",
                        "in": "query",
                        "description": "Filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[packages_installed]",
                        "in": "query",
                        "description": "Filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[packages_updatable]",
                        "in": "query",
                        "description": "Filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[system_profile][sap_system]",
                        "in": "query",
                        "description": "Filter only SAP systems",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[system_profile][sap_sids][in]",
                        "in": "query",
                        "description": "Filter systems by their SAP SIDs",
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    {
                        "name": "filter[system_profile][ansible]",
                        "in": "query",
                        "description": "Filter systems by ansible",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[system_profile][ansible][controller_version]",
                        "in": "query",
                        "description": "Filter systems by ansible version",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[system_profile][mssql]",
                        "in": "query",
                        "description": "Filter systems by mssql version",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[system_profile][mssql][version]",
                        "in": "query",
                        "description": "Filter systems by mssql version",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[osname]",
                        "in": "query",
                        "description": "Filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[osminor]",
                        "in": "query",
                        "description": "Filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[osmajor]",
                        "in": "query",
                        "description": "Filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[os]",
                        "in": "query",
                        "description": "Filter OS version",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Tag filter",
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/controllers.SystemInlineItem"
                                    }
                                }
                            },
                            "text/csv": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/controllers.SystemInlineItem"
                                    }
                                }
                            }
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            },
                            "text/csv": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            },
                            "text/csv": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "RhIdentity": []
                    }
                ]
            }
        },
...
        "/api/patch/v1/systems/{inventory_id}": {
            "get": {
                "summary": "Show me details about a system by given inventory id",
                "description": "Show me details about a system by given inventory id",
                "operationId": "detailSystem",
                "parameters": [
                    {
                        "name": "inventory_id",
                        "in": "path",
                        "description": "Inventory ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/controllers.SystemDetailResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "RhIdentity": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete system by inventory id",
                "description": "Delete system by inventory id",
                "operationId": "deletesystem",
                "parameters": [
                    {
                        "name": "inventory_id",
                        "in": "path",
                        "description": "Inventory ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "RhIdentity": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "controllers.SystemDetailResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/controllers.SystemItem"
                    }
                }
            },
            "controllers.SystemInlineItem": {
                "type": "object",
                "properties": {
                    "created": {
                        "type": "string"
                    },
                    "culled_timestamp": {
                        "type": "string"
                    },
                    "display_name": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "insights_id": {
                        "type": "string"
                    },
                    "last_evaluation": {
                        "type": "string"
                    },
                    "last_upload": {
                        "type": "string"
                    },
                    "os": {
                        "type": "string"
                    },
                    "os_major": {
                        "type": "string"
                    },
                    "os_minor": {
                        "type": "string"
                    },
                    "os_name": {
                        "type": "string"
                    },
                    "other_count": {
                        "type": "integer"
                    },
                    "packages_installed": {
                        "type": "integer"
                    },
                    "packages_updatable": {
                        "type": "integer"
                    },
                    "rhba_count": {
                        "type": "integer"
                    },
                    "rhea_count": {
                        "type": "integer"
                    },
                    "rhsa_count": {
                        "type": "integer"
                    },
                    "rhsm": {
                        "type": "string"
                    },
                    "stale": {
                        "type": "boolean"
                    },
                    "stale_timestamp": {
                        "type": "string"
                    },
                    "stale_warning_timestamp": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/controllers.SystemTag"
                        }
                    },
                    "third_party": {
                        "type": "boolean"
                    }
                }
            },
            "controllers.SystemItem": {
                "type": "object",
                "properties": {
                    "attributes": {
                        "$ref": "#/components/schemas/controllers.SystemItemAttributes"
                    },
                    "id": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    }
                }
            },
            "controllers.SystemItemAttributes": {
                "type": "object",
                "properties": {
                    "created": {
                        "type": "string"
                    },
                    "culled_timestamp": {
                        "type": "string"
                    },
                    "display_name": {
                        "type": "string"
                    },
                    "insights_id": {
                        "type": "string"
                    },
                    "last_evaluation": {
                        "type": "string"
                    },
                    "last_upload": {
                        "type": "string"
                    },
                    "os": {
                        "type": "string"
                    },
                    "os_major": {
                        "type": "string"
                    },
                    "os_minor": {
                        "type": "string"
                    },
                    "os_name": {
                        "type": "string"
                    },
                    "other_count": {
                        "type": "integer"
                    },
                    "packages_installed": {
                        "type": "integer"
                    },
                    "packages_updatable": {
                        "type": "integer"
                    },
                    "rhba_count": {
                        "type": "integer"
                    },
                    "rhea_count": {
                        "type": "integer"
                    },
                    "rhsa_count": {
                        "type": "integer"
                    },
                    "rhsm": {
                        "type": "string"
                    },
                    "stale": {
                        "type": "boolean"
                    },
                    "stale_timestamp": {
                        "type": "string"
                    },
                    "stale_warning_timestamp": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/controllers.SystemTag"
                        }
                    },
                    "third_party": {
                        "type": "boolean"
                    }
                }
            },
...
            "controllers.SystemTag": {
                "type": "object",
                "properties": {
                    "key": {
                        "type": "string"
                    },
                    "namespace": {
                        "type": "string"
                    },
                    "value": {
                        "type": "string"
                    }
                }
            },
...
            "utils.ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes": {
            "RhIdentity": {
                "type": "apiKey",
                "name": "x-rh-identity",
                "in": "header"
            }
        }
    }
}